requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-property==0.2.2
backoff==2.2.1
bcrypt==4.2.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.3
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
durationpy==0.9
fastapi==0.115.4
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
google-auth==2.36.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
kubernetes==31.0.0
langchain-core==0.3.15
langchain-openai==0.2.6
langsmith==0.1.142
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
numpy==2.1.3
oauthlib==3.2.2
onnxruntime==1.20.0
openai==1.54.3
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
posthog==3.7.0
protobuf==5.28.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-keycloak==4.7.2
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.41.2
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.3
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==14.0
wheel==0.43.0
wrapt==1.16.0
python-jose
langchain-google-vertexai
mammoth
markdownify
pdfplumber
langchain